refactor(sdk): Move graph.spec and ecs.spec into top level spec module (spec.graph, spec.ecs)
